externalize SVG part of timeline to own component
This also fixes the potential cycle with useEffect on the parent element size.
