+++ src/features/timeline/Timeline.tsx
@@ -1,10 +1,9 @@
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
-import { scaleBand, scaleTime } from 'd3-scale';
 import type { MutableRefObject } from 'react';
-import { useEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 
-import { TimelineElement } from '@/features/timeline/TimelineElement';
+import { TimelineSvg } from '@/features/timeline/TimelineSvg';
 import { usePersonsSearchResults } from '@/features/timeline/usePersonsSearchResults';
 
 export function Timeline(): JSX.Element {
@@ -12,22 +11,6 @@ export function Timeline(): JSX.Element {
   const persons = searchResults.data?.entities ?? [];
 
   const parent = useRef() as MutableRefObject<HTMLDivElement | null>;
-  const [svgViewBox, setSvgViewBox] = useState('0 0 0 0');
-  const [svgWidth, setSvgWidth] = useState(0);
-  const [svgHeight, setSvgHeight] = useState(0);
-
-  // XXX
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  useEffect(() => {
-    if (!parent.current) return;
-
-    const w = parent.current.clientWidth;
-    const h = parent.current.clientHeight;
-
-    setSvgWidth(w);
-    setSvgHeight(h);
-    setSvgViewBox(`0 0 ${w} ${h}`);
-  });
 
   if (searchResults.isLoading) {
     return (
@@ -63,27 +46,9 @@ export function Timeline(): JSX.Element {
     );
   }
 
-  const scaleX = scaleTime()
-    .domain([new Date(Date.UTC(1800, 0, 1)), new Date(Date.UTC(2020, 11, 31))])
-    .range([50, svgWidth - 50]);
-  const scaleY = scaleBand()
-    .domain(
-      persons.map((d) => {
-        return d.id;
-      }),
-    )
-    .range([50, svgHeight - 50])
-    .paddingInner(0.2);
-
   return (
     <div className="timeline-wrapper" ref={parent}>
-      <svg width="100%" height="100%" viewBox={svgViewBox}>
-        {persons.map((person) => {
-          return (
-            <TimelineElement key={person.id} scaleX={scaleX} scaleY={scaleY} person={person} />
-          );
-        })}
-      </svg>
+      <TimelineSvg parentRef={parent} persons={persons} />
     </div>
   );
 }
